+++ fruit.py
@@ -1,28 +1,23 @@
 import random
-from math import floor
-
-import pygame as pg
-from pygame.color import THECOLORS as COLORS
 from pygame.math import Vector2 as V2
 from collections import namedtuple
-from utilities import generate_random_color
-import screen as sc
+from utilities import generate_random_color, M_HEIGHT, M_WIDTH
 
 
+# Global Variables
 PROBA_APPARITION_FRUIT = 0.05
 NB_MAX_FRUIT = 40
 RAYON_FRUIT_MIN = 5
 RAYON_FRUIT_MAX = 12
 
-
-# génération fruit
 Fruit = namedtuple("Fruit", ["xy", "color", "radius"])
 LIST_FRUITS = []
 
 
+# Fruits functions
 def generate_random_fruit_position():
     """génère une position aléatoire"""
-    x, y = random.randint(0, sc.M_WIDTH), random.randint(0, sc.M_HEIGHT)
+    x, y = random.randint(0, M_WIDTH), random.randint(0, M_HEIGHT)
     return V2(x, y)
 
 
@@ -43,24 +38,11 @@ def generate_fruit():
         LIST_FRUITS.append(Fruit(xy, color, radius))
 
 
-# affichage fruits
-
-
-def draw_fruits(screen, position):
-    """affiche les fruits"""
-    for f in LIST_FRUITS:
-        center = f.xy - position + sc.SCREEN_CENTER
-        pg.draw.circle(screen, f.color, center, f.radius)
-
-
-# Manger fruit
-
-
 def eat_fruit(position, size) -> int:
     """si le fruit est assez proche, le mange. Renvoie la nouvelle taille après absorbation d'un ou plusieurs fruits"""
     for f in LIST_FRUITS:
         if (position - f.xy).length() < size:
             # formule pour ajouter à l'air du blob l'air du fruit
-            size = (size**3 + f.radius**3) ** (1 / 3)
+            size = (size**2 + f.radius**2) ** (1 / 2)
             del LIST_FRUITS[LIST_FRUITS.index(f)]
     return size
